Move `config/es` into `lib/es`
