+++ lib/es.js
@@ -2,7 +2,7 @@
  * ElasticSearch instance ready to use
  */
 var elasticsearch = require('elasticsearch');
-var cfg = require('./cfg');
+var cfg = require('./../config/cfg');
 
 
 exports = module.exports = new elasticsearch.Client({
