+++ lib/utils.js
@@ -1,5 +1,5 @@
 var Q = require('q');
-var es = require('.././config/es');
+var es = require('./es');
 
 var log = console.log;
